+++ CHANGELOG.rst
@@ -4,6 +4,9 @@ Release 2.1.8 (in development)
 * Quickhull bugfix for precision argument in degenerate cases
   (issue #3163949, fix contributed by liuhuanjim013).
 
+* Fixed issue with detecting box shapes on degenerate collision meshes
+  (fixes issue #3145104, reported by Gratis_monsta).
+
 Release 2.1.7 (23 January 2011)
 ===============================